+++ CHANGELOG.md
@@ -5,6 +5,13 @@
 All notable changes to the package will be documented here.
 
 ---
+<!-- 
+## [4.25.20] - 2025-11-06
+
+- **Added** new *constant* `TIME_ZONE_IDs` scrapped from [**IANA TZ Database**](https://en.wikipedia.org/wiki/List_of_tz_database_time_zones)
+- **Added** new *guard* `isValidTimeZoneId` to check if a string value is a valid *timezone identifier* from **IANA Database**.
+- **Updated** `Chronos` `timeZonePlugin` method `timeZone` to accept `TimeZoneIdentifier` along with *short timezone names* and *UTC offset*.
+-->
 
 ## [4.25.11] - 2025-11-06